refactor: extract telegram basic command handler
